Redesign unlock all page
