+++ source/setup/library/messaging.js
@@ -45,19 +45,6 @@ export function getLastLogin() {
     });
 }
 
-export function lockAllArchives() {
-    log.info("Sending request to background to lock all archives");
-    return new Promise((resolve, reject) => {
-        chrome.runtime.sendMessage({ type: "lock-all-archives" }, response => {
-            const { ok, error } = response;
-            if (ok) {
-                return resolve();
-            }
-            return reject(new Error(`Locking archives failed: ${error}`));
-        });
-    });
-}
-
 export function lockArchive(sourceID) {
     log.info(`Sending request to background to lock archive source: ${sourceID}`);
     return new Promise((resolve, reject) => {
